Refactor component names and added tons of documentation
