+++ app/src/main/java/com/example/gymsmart/components/Login.kt
@@ -6,7 +6,12 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 
-
+/**
+ * Login
+ *
+ * @param onGoogleSignInClick
+ * @receiver
+ */
 @Composable
 fun Login(onGoogleSignInClick: () -> Unit) {
     Button(
