+++ app/src/main/java/com/example/gymsmart/firebase/UserSession.kt
@@ -1,5 +1,10 @@
 package com.example.gymsmart.firebase
 
+/**
+ * User session
+ *
+ * @constructor Create empty User session
+ */
 object UserSession {
     var userEmail: String? = null
     var userName: String? = null
